[xUnit 3] Convert Akka.Persistence.TestKit and Akka.Persistence.TestKit.Tests
