+++ Akka.sln
@@ -281,6 +281,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Akka.TestKit.Xunit.Tests", 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Akka.Tests.Shared.Internals.Xunit3", "src\core\Akka.Tests.Shared.Internals.Xunit3\Akka.Tests.Shared.Internals.Xunit3.csproj", "{24DD484D-0F15-4D13-B603-F2E4A901CDBC}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Akka.Persistence.TestKit.Xunit", "src\core\Akka.Persistence.TestKit.Xunit\Akka.Persistence.TestKit.Xunit.csproj", "{164DFDEA-609E-43F6-AE5F-1A013A990A9F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -1332,6 +1334,18 @@ Global
 		{24DD484D-0F15-4D13-B603-F2E4A901CDBC}.Release|x64.Build.0 = Release|Any CPU
 		{24DD484D-0F15-4D13-B603-F2E4A901CDBC}.Release|x86.ActiveCfg = Release|Any CPU
 		{24DD484D-0F15-4D13-B603-F2E4A901CDBC}.Release|x86.Build.0 = Release|Any CPU
+		{164DFDEA-609E-43F6-AE5F-1A013A990A9F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{164DFDEA-609E-43F6-AE5F-1A013A990A9F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{164DFDEA-609E-43F6-AE5F-1A013A990A9F}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{164DFDEA-609E-43F6-AE5F-1A013A990A9F}.Debug|x64.Build.0 = Debug|Any CPU
+		{164DFDEA-609E-43F6-AE5F-1A013A990A9F}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{164DFDEA-609E-43F6-AE5F-1A013A990A9F}.Debug|x86.Build.0 = Debug|Any CPU
+		{164DFDEA-609E-43F6-AE5F-1A013A990A9F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{164DFDEA-609E-43F6-AE5F-1A013A990A9F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{164DFDEA-609E-43F6-AE5F-1A013A990A9F}.Release|x64.ActiveCfg = Release|Any CPU
+		{164DFDEA-609E-43F6-AE5F-1A013A990A9F}.Release|x64.Build.0 = Release|Any CPU
+		{164DFDEA-609E-43F6-AE5F-1A013A990A9F}.Release|x86.ActiveCfg = Release|Any CPU
+		{164DFDEA-609E-43F6-AE5F-1A013A990A9F}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -1456,6 +1470,7 @@ Global
 		{95017C99-E960-44E5-83AD-BF21461DF06F} = {7625FD95-4B2C-4A5B-BDD5-94B1493FAC8E}
 		{F80F41E6-E5C7-4C92-B1CF-42539ECFBE68} = {7625FD95-4B2C-4A5B-BDD5-94B1493FAC8E}
 		{24DD484D-0F15-4D13-B603-F2E4A901CDBC} = {01167D3C-49C4-4CDE-9787-C176D139ACDD}
+		{164DFDEA-609E-43F6-AE5F-1A013A990A9F} = {01167D3C-49C4-4CDE-9787-C176D139ACDD}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {03AD8E21-7507-4E68-A4E9-F4A7E7273164}
