+++ src/core/Akka.Persistence.TestKit.Tests/Bug4762FixSpec.cs
@@ -5,15 +5,11 @@
 // </copyright>
 //-----------------------------------------------------------------------
 
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using Akka.Actor;
 using Akka.Event;
 using Xunit;
-using Xunit.Abstractions;
 
 namespace Akka.Persistence.TestKit.Tests
 {
