+++ src/core/Akka.Persistence.TestKit.Xunit/GlobalUsings.cs
@@ -1,0 +1,8 @@
+﻿// -----------------------------------------------------------------------
+//  <copyright file="GlobalUsings.cs" company="Akka.NET Project">
+//      Copyright (C) 2009-2022 Lightbend Inc. <http://www.lightbend.com>
+//      Copyright (C) 2013-2026 .NET Foundation <https://github.com/akkadotnet/akka.net>
+//  </copyright>
+// -----------------------------------------------------------------------
+
+global using XTestKit = Akka.TestKit.Xunit.TestKit;
